% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_bc}
\alias{apply_bc}
\title{Apply boundary conditions}
\usage{
apply_bc(nodes, node_ids, constraints)
}
\arguments{
\item{nodes}{Nodes data frame}

\item{node_ids}{Vector of node IDs}

\item{constraints}{List of constraints to apply}
}
\description{
Apply boundary conditions
}
