% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness_matrices.R
\name{apply_boundary_conditions}
\alias{apply_boundary_conditions}
\title{Apply boundary conditions to stiffness matrix}
\usage{
apply_boundary_conditions(K, constrained_dof)
}
\arguments{
\item{K}{Global stiffness matrix}

\item{constrained_dof}{Degrees of freedom to constrain}
}
\description{
Apply boundary conditions to stiffness matrix
}
