% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_matrices.R
\name{assemble_global_mass}
\alias{assemble_global_mass}
\title{Assemble global mass matrix}
\usage{
assemble_global_mass(elements, connectivity, total_dof)
}
\arguments{
\item{elements}{List of elements}

\item{connectivity}{Connectivity matrix}

\item{total_dof}{Total degrees of freedom}
}
\description{
Assemble global mass matrix
}
