% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness_matrices.R
\name{assemble_global_stiffness}
\alias{assemble_global_stiffness}
\title{Assemble global stiffness matrix}
\usage{
assemble_global_stiffness(elements, connectivity, total_dof)
}
\arguments{
\item{elements}{List of elements}

\item{connectivity}{Connectivity matrix (n x 2)}

\item{total_dof}{Total degrees of freedom}
}
\description{
Assemble global stiffness matrix
}
