% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_types.R
\name{BeamElement}
\alias{BeamElement}
\title{Beam element (Euler-Bernoulli)}
\usage{
BeamElement(L, E, I, A, rho)
}
\arguments{
\item{L}{Length (m)}

\item{E}{Young's modulus (Pa)}

\item{I}{Moment of inertia (m4)}

\item{A}{Cross-sectional area (m2)}

\item{rho}{Density (kg/m3)}
}
\description{
Beam element (Euler-Bernoulli)
}
