% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_material}
\alias{check_material}
\title{Check material properties}
\usage{
check_material(E, nu, rho, fy = NULL)
}
\arguments{
\item{E}{Young's modulus (Pa)}

\item{nu}{Poisson's ratio}

\item{rho}{Density (kg/m3)}

\item{fy}{Yield strength (Pa)}
}
\description{
Check material properties
}
