% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_models.R
\name{create_2D_frame}
\alias{create_2D_frame}
\title{Create a simple 2D frame building}
\usage{
create_2D_frame(n_stories = 3, story_height = 3, bay_width = 5, n_bays = 3)
}
\arguments{
\item{n_stories}{Number of stories}

\item{story_height}{Height per story (m)}

\item{bay_width}{Bay width (m)}

\item{n_bays}{Number of bays}
}
\description{
Create a simple 2D frame building
}
