% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_models.R
\name{create_3D_building}
\alias{create_3D_building}
\title{Create a 3D building model}
\usage{
create_3D_building(
  n_stories = 3,
  plan_width = 10,
  plan_depth = 10,
  story_height = 3
)
}
\description{
Create a 3D building model
}
