% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismic_forces.R
\name{generate_earthquake}
\alias{generate_earthquake}
\title{Generate earthquake acceleration record}
\usage{
generate_earthquake(duration = 30, dt = 0.01, type = "el_centro", PGA = 0.3)
}
\arguments{
\item{duration}{Total duration (s)}

\item{dt}{Time step (s)}

\item{type}{Type of record: "el_centro", "kobe", "sine", "white_noise"}

\item{PGA}{Peak Ground Acceleration (g)}
}
\description{
Generate earthquake acceleration record
}
