% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{newmark_beta}
\alias{newmark_beta}
\title{Solve time history analysis (Newmark-beta method)}
\usage{
newmark_beta(M, C, K, F_time, dt, beta = 0.25, gamma = 0.5)
}
\arguments{
\item{M}{Mass matrix}

\item{C}{Damping matrix}

\item{K}{Stiffness matrix}

\item{F_time}{Time-varying force matrix}

\item{dt}{Time step}

\item{beta}{Newmark-beta parameter}

\item{gamma}{Newmark-gamma parameter}
}
\description{
Solve time history analysis (Newmark-beta method)
}
