% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mode_shape}
\alias{plot_mode_shape}
\title{Plot mode shapes}
\usage{
plot_mode_shape(nodes, modes, mode_number = 1, scale_factor = 10)
}
\arguments{
\item{nodes}{Nodes data frame}

\item{modes}{Mode shapes matrix}

\item{mode_number}{Which mode to plot}

\item{scale_factor}{Scale factor for deformation}
}
\description{
Plot mode shapes
}
