% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_time_history}
\alias{plot_time_history}
\title{Plot time history response}
\usage{
plot_time_history(
  time,
  response,
  node_id = 1,
  dof = 1,
  title = "Time History Response"
)
}
\arguments{
\item{time}{Time vector}

\item{response}{Response vector (displacement, velocity, or acceleration)}

\item{node_id}{Node ID to plot}

\item{dof}{Degree of freedom (1-6)}
}
\description{
Plot time history response
}
