% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damping_matrices.R
\name{rayleigh_damping}
\alias{rayleigh_damping}
\title{Create Rayleigh damping matrix}
\usage{
rayleigh_damping(K, M, xi = 0.05, omega1 = NULL, omega2 = NULL)
}
\arguments{
\item{K}{Stiffness matrix}

\item{M}{Mass matrix}

\item{xi}{Damping ratio (e.g., 0.05 for 5%)}

\item{omega1}{First natural frequency (rad/s)}

\item{omega2}{Second natural frequency (rad/s)}
}
\description{
Create Rayleigh damping matrix
}
