% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismic_forces.R
\name{response_spectrum_analysis}
\alias{response_spectrum_analysis}
\title{Calculate seismic forces using response spectrum}
\usage{
response_spectrum_analysis(M, K, spectrum, directions = c(1, 0, 0))
}
\arguments{
\item{M}{Mass matrix}

\item{K}{Stiffness matrix}

\item{spectrum}{Response spectrum function}
}
\description{
Calculate seismic forces using response spectrum
}
