% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_types.R
\name{ShearWallElement}
\alias{ShearWallElement}
\title{Shear wall element}
\usage{
ShearWallElement(height, width, thickness, E, rho, nu = 0.2)
}
\arguments{
\item{height}{Wall height (m)}

\item{width}{Wall width (m)}

\item{thickness}{Wall thickness (m)}
}
\description{
Shear wall element
}
