% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_modes}
\alias{solve_modes}
\title{Solve modal analysis}
\usage{
solve_modes(M, K, n_modes = 10)
}
\arguments{
\item{M}{Mass matrix}

\item{K}{Stiffness matrix}

\item{n_modes}{Number of modes to compute}
}
\description{
Solve modal analysis
}
