% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_static}
\alias{solve_static}
\title{Solve static analysis}
\usage{
solve_static(K, F, constrained_dof)
}
\arguments{
\item{K}{Global stiffness matrix}

\item{F}{Force vector}

\item{constrained_dof}{Constrained degrees of freedom}
}
\description{
Solve static analysis
}
