% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_building}
\alias{visualize_building}
\title{Create building visualization}
\usage{
visualize_building(
  nodes,
  elements,
  connectivity,
  deformation = NULL,
  scale = 10
)
}
\description{
Create building visualization
}
